+++ erp_back/src/modules/invoice/model.rs
@@ -35,7 +35,7 @@ pub struct SaleCondition {
 pub struct LineItem {
     pub product: LineProduct,
     pub unit_cost: Decimal,
-    pub tax: i8,
+    pub tax: Decimal,
     pub quantity: i32,
 }
